removed video

+++ src/_Utility/listwrappers.h
@@ -103,11 +103,11 @@ class IntItemIDWrapper
 public:
     IntItemIDWrapper() {}
     IntItemIDWrapper(T t) : m_t(t) {}
-    IntItemIDWrapper(ItemID pitem) { m_t = ((int)pitem == ZeroValue) ? 0 : (T)pitem; }
+    IntItemIDWrapper(ItemID pitem) { m_t = ((intptr_t)pitem == (intptr_t)ZeroValue) ? 0 : (T)(intptr_t)pitem; }
 
     int operator = (T t) { m_t = t; }
     operator T () { return m_t; }
-    operator ItemID () { return (ItemID)((m_t == 0) ? (T)ZeroValue : m_t); };
+    operator ItemID () { return (ItemID)(intptr_t)((m_t == 0) ? (T)ZeroValue : m_t); };
 
     bool operator == (IntItemIDWrapper<T> t) const { return m_t == t; }
     bool operator <= (IntItemIDWrapper<T> t) const { return m_t <= t; }
